Move non-LLPC declarations and code out of llpc directory  (step 1).
llpc.h seperation. llpc-sepecific code in llpc.h and others used by the
XGL driver without building LLPC in compiler/include/vkgcDefs.h.
